Fix Scientist kill progression on assisting

+++ src/game/server/infclass/classes/humans/human.h
@@ -107,7 +107,7 @@ private:
 
 	int m_BonusTick = 0;
 	int m_ResetKillsTick = 0;
-	int m_KillsProgression = 0;
+	float m_KillsProgression = 0;
 	int m_TurretCount = 0;
 	int m_BroadcastWhiteHoleReady; // used to broadcast "WhiteHole ready" for a short period of time
 	int m_PositionLockTicksRemaining = 0;
